+++ src/main/java/com/project/itbar/repos/UserRepo.java
@@ -3,6 +3,8 @@ package com.project.itbar.repos;
 import com.project.itbar.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepo extends JpaRepository<User, Long> {
+import java.math.BigInteger;
+
+public interface UserRepo extends JpaRepository<User, BigInteger> {
     User findByUsername(String username);
 }
